+++ src/Conversion/ONNXToTOSA/DialectBuilder.hpp
@@ -38,7 +38,7 @@ struct TosaBuilder : DialectBuilder {
   virtual ~TosaBuilder() {}
 
   mlir::Value reshape(mlir::Value &value, llvm::ArrayRef<int64_t> shape);
-  mlir::Value transpose(mlir::Value &value, llvm::ArrayRef<int64_t> perm);
+  mlir::Value transpose(mlir::Value &value, llvm::ArrayRef<int32_t> perm);
   mlir::Value slice(mlir::Value &inputConst, llvm::ArrayRef<int64_t> size,
       llvm::ArrayRef<int64_t> start);
